% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_wrap.R
\name{estimate_ensmallen_grad}
\alias{estimate_ensmallen_grad}
\title{estimate_ensmallen_grad}
\usage{
estimate_ensmallen_grad(
  j,
  df_l,
  X_l,
  h1 = NULL,
  use_sparse = FALSE,
  tau0 = 0,
  tau1 = Inf,
  sce = 2.1,
  tol = 1e-08,
  max_iter = 1000,
  init
)
}
\value{
estimation stuff
}
\description{
estimate_ensmallen_grad
}
