% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estNP_wrap.R
\name{estimate_NP}
\alias{estimate_NP}
\title{estimate_NP}
\usage{
estimate_NP(
  j,
  df_l,
  X_l,
  h1 = NULL,
  h2 = NULL,
  d = 0,
  tau0 = 0,
  tau1 = Inf,
  tol = 1e-08,
  max_iter = 1000,
  init
)
}
\value{
estimation of beta
}
\description{
estimate_NP
}
