% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_est_wrap.R
\name{estimate_survivor_grad}
\alias{estimate_survivor_grad}
\title{estimate_survivor_grad
L-BFGS for survivor model (Section 4)}
\usage{
estimate_survivor_grad(
  df_s,
  X_s,
  h1 = NULL,
  use_sparse = FALSE,
  tau0 = 0,
  tau = Inf,
  tol = 1e-08,
  max_iter = 1000,
  init
)
}
\description{
estimate_survivor_grad
L-BFGS for survivor model (Section 4)
}
