% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_wrapper.R
\name{mu_hat}
\alias{mu_hat}
\title{\code{mu_hat}: baseline mean estimator \hat{\mu}(t,s) for any scenario}
\usage{
mu_hat(j, t, s, h, par, X, Y, deltaPi, S, Z, scenario)
}
\arguments{
\item{j}{integer {1,2}}

\item{t, s}{evaluation time t and landmark s (same length if vectors)}

\item{h}{bandwidth}

\item{par}{numeric vector: c(beta_j, theta_j)}

\item{X, Y}{design matrix (p x n) and response (n)}

\item{deltaPi}{length-n integer vector with values in {1,2}}

\item{scenario}{numeric, like 1.1, 1.2, 2.1, 2.2}

\item{A, Z}{length-n subject-level s and time vectors}
}
\description{
\code{mu_hat}: baseline mean estimator \hat{\mu}(t,s) for any scenario
}
