% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_wrapper.R
\name{mu_hat_fast}
\alias{mu_hat_fast}
\title{\code{mu_hat_fast}: same as \code{mu_hat} but filters deltaPi==j in R for speed}
\usage{
mu_hat_fast(
  j,
  t,
  s,
  h,
  par,
  X,
  Y,
  deltaPi,
  S,
  Z,
  scenario,
  tau_norm,
  center = 0
)
}
\arguments{
\item{j}{integer {1,2}}

\item{t, s}{evaluation time t and landmark s (scalars; vectorize via mapply)}

\item{h}{bandwidth}

\item{par}{numeric vector: c(beta_j, theta_j)}

\item{X, Y}{design matrix (p x n) and response (n)}

\item{deltaPi}{length-n integer vector with values in {1,2}}

\item{S, Z}{length-n visit time and event/censoring time vectors}

\item{scenario}{numeric, like 1.1, 1.2, 2.1, 2.2}
}
\description{
\code{mu_hat_fast}: same as \code{mu_hat} but filters deltaPi==j in R for speed
}
