% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muNP_wrapper.R
\name{muNP_hat}
\alias{muNP_hat}
\title{\code{muNP_hat}: baseline mean estimator \hat{\mu}(t,s)}
\usage{
muNP_hat(j, t, s, h1, h2, par, X, Y, deltaPi, S, Z)
}
\arguments{
\item{j}{integer {1,2}}

\item{t, s}{evaluation time t and landmark s (same length if vectors)}

\item{h1}{bandwidth1}

\item{h2}{bandwidth2}

\item{par}{numeric vector: c(beta_j)}

\item{X, Y}{design matrix (p x n) and response (n)}

\item{deltaPi}{length-n integer vector with values in {1,2}}

\item{A, Z}{length-n subject-level s and time vectors}
}
\description{
\code{muNP_hat}: baseline mean estimator \hat{\mu}(t,s)
}
