% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajecEstm-package.R
\docType{package}
\name{TrajecEstm-package}
\alias{TrajecEstm}
\alias{TrajecEstm-package}
\title{TrajecEstm: Trajectory Estimation}
\description{
Tools for trajectory estimation and survival-related estimation using kernel-based methods and gradient-based optimization.
}
\author{
\strong{Maintainer}: Ziyi Wang \email{ziyi.w@columbia.edu}

}
\keyword{internal}
